requirements.txt
streamlit
requests
bs4
beautifulsoup4
psycopg2-binary
tqdm
numpy
